requirements.txt
Flask==3.0.0
python-telegram-bot==20.6
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0
gunicorn==21.2.0
